% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWaterVolumes.R
\name{checkWaterVolumes}
\alias{checkWaterVolumes}
\title{Check infiltrated water volumen}
\usage{
checkWaterVolumes(sm_mod_data, whichCheck, inf_time, cell_volume,
  infWater_timestep, ts_check = 1, plotting = FALSE)
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
