% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_SGpillar.R
\name{correct_SGpillar}
\alias{correct_SGpillar}
\title{Correct gravity component grid for the SG pillar}
\usage{
correct_SGpillar(gravity_comp3d, Pillar_x = NA, Pillar_y = NA,
  Pillar_z = NA, correct_radius = NA, correct_depth = NA, SG_X = SG_x,
  SG_Y = SG_y, grid_discretization = NA)
}
\arguments{
\item{Pillar_x, y, z}{Numeric, coordinate of the SG pillar (x and y).}

\item{correct_radius}{Numeric, radius of the SG pillar.}

\item{correct_depth}{Numeric, vertical extent of the SG pillar.}

\item{grid_discretization}{test}

\item{gravity_gcomp3d}{Data.frame, of the gravity component grid containing columns x, y, z, zgrid and gcomp.}

\item{SG_x, y}{Numeric, coordinate of the SG (x and y).}
}
\value{
A data.frame of the corrected gravity component grid is return.
}
\description{
Removes gravity components at location of the SG pillar.
 It can correct for both rectangular or circular shaped pillars, 
 depending on what data is supplied.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
