% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_WaterIntesityGrid.R
\name{create_WaterIntensityGrid}
\alias{create_WaterIntensityGrid}
\title{Create an interpolated grid of the superficial water intensity distribution}
\usage{
create_WaterIntensityGrid(input_file, intp_method, surface, zerosBorder,
  spat_col, dat_col, UTM_gridcenter_x = NA, UTM_gridcenter_y = NA,
  input_dir = dir_input, output_dir = dir_output, ...)
}
\arguments{
\item{test}{test}
}
\value{
A data.frame, containing spatial information (x, y) and a scale factor of the water intensity for each grid cell.
}
\description{
Based on the supplied model domain and measured (or constructed) water intensities,
 a surface grid is interpolated for further usage as infiltration model input.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
