% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_model_3d_2macro.R
\name{inf_model_3d_2macro}
\alias{inf_model_3d_2macro}
\title{Hydro-gravimetric model (incl. lateral flow), for 2 different macro pore layers}
\usage{
inf_model_3d_2macro(param_vec)
}
\arguments{
\item{paramvec}{Numeric vector, proving input in this EXACT ORDER of the parameter values: saturation deficit of 1st process (macro pores), saturation deficit of 2nd process, total vertical thickness (extent) of 1st process (macropores), scaling factor (0 to 1) defining the relatinship between lateral and vertical flow after saturation of a cell.}

\item{test}{...}
}
\description{
3D distribution algorithm, modeling the soil infiltration of water from a sprinkling experiment
}
\details{
The function was written to be easily usable together with optimization tools.
It thus  needs a configfile.rdata-file with the following columns, each consisting of one single value:
dir_input, dir_output, precip_time, IntensityDistributionution, water_vol_min, gcompfile, gravityObs, mb_permitted_error.
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
