% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gravity_reponses.R
\name{plot_gravity_responses}
\alias{plot_gravity_responses}
\title{Plot gravity response of observed and modeled gravity signal}
\usage{
plot_gravity_responses(gravity_obs, gravity_mod, input_dir = dir_input,
  output_dir = dir_output, dat_tsf = 7, ...)
}
\arguments{
\item{gravity_obs}{Character string with the filename of the gravity observation data to be used.}

\item{gravity_mod}{Character string with the filename of the modeled gravity response data.}

\item{input_dir}{Character string, containing the path to input file "gravity_obs".}

\item{output_dir}{Character string, containing the path to the directory where to save the plot.}
}
\value{
Returns NULL as object. It creates a plot of all input time series.
This plot is shown on screen as well as saved to the specified output folder.
}
\description{
Plot gravity time series obtained through measurements during the sprinkling
experiment and modeling of infiltration of sprinkled water.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
