% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tempWaterDistribution.R
\name{plot_tempWaterDistribution}
\alias{plot_tempWaterDistribution}
\title{test}
\usage{
plot_tempWaterDistribution(sm_mod_data, plotvar)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
