% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transects_2d.R
\name{plot_transects_2d}
\alias{plot_transects_2d}
\title{Plot 2d transect of modeled infiltration waster distribution}
\usage{
plot_transects_2d(soilmoisture_mod, plot_int, y_pos, vert_limit = NA,
  output_dir = dir_output, ...)
}
\arguments{
\item{soilmoisture_mod}{Character string with the filename of the modeled soil moisture data.}

\item{output_dir}{Character string, containing the path to the directory where to save the plot.}

\item{plot_ts}{Vector of one or more time stamps at which plotting is desired.
They have to be supplied in the unit "XX".}

\item{input_dir}{Character string, containing the path to input file "gravity_obs".}
}
\value{
Returns NULL as object. It creates a plot of all input time series.
This plot is shown on screen as well as saved to the specified output folder.
}
\description{
Plot a transect in 2d of the 3d modeled infiltration water distribution.
One or more specific time steps of the total sprinkling experiment duration can be chosen.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
