% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotInfiltrationProfile}
\alias{plotInfiltrationProfile}
\title{plot irrigated cut profile at certain time step}
\usage{
plotInfiltrationProfile(filein, tstep, cutrel, plotvar = "x",
  showlegend = T)
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
