% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_conversion.R
\name{run_model_conversion}
\alias{run_model_conversion}
\title{Run infiltration model in conversion mode}
\usage{
run_model_conversion(dtheta_macro_start = dtheta_macro_start,
  dtheta_macro2_start = dtheta_macro2_start, mdepth_start = mdepth_start,
  mdepth2_start = mdepth2_start, dtheta_other_start = dtheta_other_start,
  latflow_fac_start = latflow_fac_start, output_dir, ...)
}
\arguments{
\item{test}{test}
}
\value{
A list (?) of the model parameters and model input is returned.
}
\description{
The infiltration model is run one in conversion mode.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
